+++ auth/auth_user.go
@@ -1,0 +1,7 @@
+package auth
+
+type AuthUser interface {
+	GetID() string
+	GetEmail() string
+	GetHashedPassword() string
+}