+++ http/http_engine.go
@@ -1,7 +1,7 @@
 package http
 
 type HandlerFunc func(ctx Context)
-type MiddlewareFunc func(ctx Context)
+type MiddlewareFunc func(next HandlerFunc) HandlerFunc
 
 type HTTPEngine interface {
 	Init(port int) error