+++ utils/generate_from_password.go
@@ -1,0 +1,14 @@
+package utils
+
+import (
+	"fmt"
+	"golang.org/x/crypto/bcrypt"
+)
+
+func GenerateFromPassword(password string) (string, error) {
+	hashed, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+	if err != nil {
+		return "", fmt.Errorf("failed to hash password")
+	}
+	return string(hashed), nil
+}